Use pass_config in all commands, except the main one.
